spapi_probe/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
google-cloud-bigquery==3.17.1
requests-aws4auth==1.2.3
pytz==2023.3.post1
gunicorn==21.2.0
